ThemeSetConversion implemented, ThemeSetServiceImpl null parameter restrictions

+++ src/main/java/cz/muni/fi/PA165/service/ThemeSetServiceImpl.java
@@ -5,6 +5,7 @@
 package cz.muni.fi.PA165.service;
 
 import cz.muni.fi.PA165.dao.ThemeSetDao;
+import cz.muni.fi.PA165.daoDtoConversion.ThemeSetConversion;
 import cz.muni.fi.PA165.dto.ThemeSetDto;
 import cz.muni.fi.PA165.entity.ThemeSet;
 import java.math.BigDecimal;
@@ -31,7 +32,10 @@ public class ThemeSetServiceImpl implements ThemeSetService{
     }
     
     public void create(ThemeSetDto setDto) {
-        ThemeSet ts = setDto.createEntity();
+        if (setDto == null){
+            throw new IllegalArgumentException("DTO object cannot be NULL");
+        }
+        ThemeSet ts = ThemeSetConversion.convertToEntity(setDto);
         themeSetDao.create(ts);
     }
 
@@ -45,6 +49,9 @@ public class ThemeSetServiceImpl implements ThemeSetService{
     }
 
     public List<ThemeSetDto> findByPrice(BigDecimal price) {
+        if (price == null){
+            throw new IllegalArgumentException("Parameter price cannot be NULL");
+        }
         List<ThemeSet> tsList = themeSetDao.findByPrice(price);
         List<ThemeSetDto> themeSetDtoList = new ArrayList<ThemeSetDto>();
         for (ThemeSet ts : tsList){
@@ -54,15 +61,24 @@ public class ThemeSetServiceImpl implements ThemeSetService{
     }
 
     public void update(ThemeSetDto setDto) {
-        ThemeSet ts = setDto.createEntity();
+        if (setDto == null){
+            throw new IllegalArgumentException("DTO object cannot be NULL");
+        }
+        ThemeSet ts = ThemeSetConversion.convertToEntity(setDto);
         themeSetDao.update(ts);
     }
 
     public void delete(Long id) {
+        if (id == null){
+            throw new IllegalArgumentException("ID cannot be NULL");
+        }
         themeSetDao.delete(id);
     }
 
     public ThemeSetDto findById(Long id) {
+        if (id == null){
+            throw new IllegalArgumentException("ID cannot be NULL");
+        }
         ThemeSet entity = themeSetDao.findById(id);
         return entity.createDto();
     }
